fix positioning of create new article bottom toolbar button
