+++ scripts/apps/authoring/views/opened-articles.html
@@ -25,7 +25,7 @@
         </li>
     </ul>
 
-    <div class="opened-articles-bar__create-new dropdown dropdown--dropleft strict" dropdown ng-if="privileges.archive">
+    <div class="opened-articles-bar__create-new" ng-if="privileges.archive">
         <sd-content-create custom-button="createNewButton"></sd-content-create>
     </div>
 </div>
